refactor: moves types as close to where they define the interfaces as possible
