+++ src/codeowners/generate.test.ts
@@ -2,11 +2,11 @@ import { deepEqual, equal } from "node:assert/strict";
 import { readFileSync } from "node:fs";
 import { EOL } from "node:os";
 import { describe, it } from "node:test";
-import type { ITeamMap } from "../../types/team-map.js";
-import generateCodeOwners from "./generate.js";
-import { parse } from "../virtual-code-owners/parse.js";
 import readTeamMap from "../team-map/read.js";
+import type { ITeamMap } from "../team-map/team-map.js";
+import { parse } from "../virtual-code-owners/parse.js";
 import readVirtualCodeOwners from "../virtual-code-owners/read.js";
+import generateCodeOwners from "./generate.js";
 
 export function generateCodeOwnersFromString(
   pCodeOwnersFileAsString: string,
