+++ src/virtual-code-owners/cst.d.ts
