--- types/cli.d.ts
@@ -1,6 +1,0 @@
-import { IOptions } from "./main.js";
-
-export interface ICommandLineOptions extends IOptions {
-  help: boolean;
-  version: boolean;
-}
