--- types/generate-codeowners.d.ts
@@ -1,6 +1,0 @@
-import type { ITeamMap, IVirtualCodeOwnersCST } from "./types.js";
-export default function generateCodeOwners(
-  pVirtualCodeOwners: IVirtualCodeOwnersCST,
-  pTeamMap: ITeamMap,
-  pGeneratedWarning?: string,
-): string;
