--- types/generate-labeler-yml.d.ts
@@ -1,7 +1,0 @@
-import type { ITeamMap, IVirtualCodeOwnersCST } from "./types.js";
-
-export default function generateLabelerYml(
-  pCodeOwners: IVirtualCodeOwnersCST,
-  pTeamMap: ITeamMap,
-  pGeneratedWarning?: string,
-): string;
