--- types/main.d.ts
@@ -1,8 +1,0 @@
-export interface IOptions {
-  virtualCodeOwners: string;
-  virtualTeams: string;
-  codeOwners: string;
-  emitLabeler: boolean;
-  labelerLocation: string;
-  dryRun: boolean;
-}
