--- types/types.d.ts
@@ -1,5 +1,0 @@
-export * from "./main.js";
-export * from "./cli.js";
-export * from "./virtual-code-owners-cst.js";
-export * from "./team-map.js";
-export * from "./anomaly.js";
